requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==7.0.0
DateTime==5.5
idna==3.7
pytz==2024.1
requests==2.31.0
urllib3==2.2.1
zope.interface==6.3
pymongo==4.7.2
python-dotenv==1.0.1
dataclasses-json==0.6.6